requirements.txt
# Core Dependencies
torch>=1.9.0
transformers>=4.20.0
soundfile>=0.12.1
numpy>=1.21.0
PyYAML>=6.0

# Audio Processing
pyaudio>=0.2.11
wave

# Optional Dependencies
pygame>=2.1.0  # For audio playback
python-dotenv>=0.19.0  # For environment variables
playsound>=1.3.0  # Alternative audio playback

# Development Dependencies
pytest>=6.0.0
pre-commit>=2.15.0

# Optional: Cloud Services
# google-cloud-speech>=2.16.0  # Google Speech-to-Text API

# Optional: For advanced STT
# whisper  # OpenAI Whisper (if using whisper instead of whisper.cpp)
# openai-whisper>=20231117  # Alternative whisper implementation

# Optional: For wake word detection
# pvporcupine>=3.0.0  # Picovoice Porcupine

# Optional: For advanced TTS
# TTS>=0.22.0  # Coqui TTS
# piper-tts  # Piper TTS

# Optional: Additional audio processing
# librosa>=0.10.0  # Advanced audio analysis
# scipy>=1.9.0  # Scientific computing for audio processing
